+++ src/main/java/ru/khasang/cachoeira/view/mainwindow/ProjectPropertiesPaneController.java
@@ -1,4 +1,4 @@
-package ru.khasang.cachoeira.view;
+package ru.khasang.cachoeira.view.mainwindow;
 
 import javafx.beans.InvalidationListener;
 import javafx.beans.WeakInvalidationListener;
@@ -8,6 +8,7 @@ import javafx.scene.control.DatePicker;
 import javafx.scene.control.TextArea;
 import javafx.scene.control.TextField;
 import javafx.scene.input.KeyCode;
+import ru.khasang.cachoeira.view.UIControl;
 
 import java.time.LocalDate;
