+++ src/main/java/ru/khasang/cachoeira/view/mainwindow/contextmenus/TaskContextMenu.java
@@ -1,4 +1,4 @@
-package ru.khasang.cachoeira.view.contextmenus;
+package ru.khasang.cachoeira.view.mainwindow.contextmenus;
 
 import javafx.collections.ObservableList;
 import javafx.scene.control.CheckMenuItem;
@@ -6,30 +6,37 @@ import javafx.scene.control.ContextMenu;
 import javafx.scene.control.Menu;
 import javafx.scene.control.MenuItem;
 import ru.khasang.cachoeira.controller.IController;
-import ru.khasang.cachoeira.model.DependentTask;
-import ru.khasang.cachoeira.model.IDependentTask;
-import ru.khasang.cachoeira.model.ITask;
-import ru.khasang.cachoeira.model.TaskDependencyType;
+import ru.khasang.cachoeira.model.*;
 import ru.khasang.cachoeira.view.UIControl;
 
 import java.util.ResourceBundle;
 
-public class ParentTaskContextMenu extends ContextMenu {
+/**
+ * Класс описывает контекстное меню всплывающее при нажатии правой кнопкой на задаче.
+ */
+public class TaskContextMenu extends ContextMenu {
     private ResourceBundle bundle = UIControl.bundle;
 
-    public ParentTaskContextMenu() {
+    public TaskContextMenu() {
+    }
+
+    public TaskContextMenu(IController controller, ITask task) {
+        initMenus(controller, task);
     }
 
     public void initMenus(IController controller, ITask task) {
         this.getItems().clear();
+        Menu assignResourceMenu = new Menu(bundle.getString("assign_resource"));
         Menu assignDependencyTask = new Menu(bundle.getString("assign_predecessor"));
+        MenuItem removeTaskMenuItem = new MenuItem(bundle.getString("remove_task"));
 
-        this.getItems().addAll(assignDependencyTask);  //заполняем меню
+        removeTaskMenuItem.setOnAction(event -> controller.handleRemoveTask(task));
+        this.getItems().addAll(assignResourceMenu, assignDependencyTask, removeTaskMenuItem);  //заполняем меню
 
-        this.setOnShowing(event -> refreshDependencyTaskMenu(
-                assignDependencyTask.getItems(),
-                task,
-                controller.getProject().getTaskList()));
+        this.setOnShowing(event -> {
+            refreshResourceMenu(assignResourceMenu.getItems(), task, controller.getProject().getResourceList());
+            refreshDependencyTaskMenu(assignDependencyTask.getItems(), task, controller.getProject().getTaskList());
+        });
     }
 
     private void refreshDependencyTaskMenu(ObservableList<MenuItem> menuItemsList,
@@ -90,6 +97,27 @@ public class ParentTaskContextMenu extends ContextMenu {
                                 .filter(menuItem -> menuItem.getText().equals(childTask.getTask().getName()))
                                 .forEach(menuItem -> menuItem.setDisable(true));
                     });
+        }
+    }
+
+    private void refreshResourceMenu(ObservableList<MenuItem> menuItemList,
+                                     ITask task,
+                                     ObservableList<IResource> resourceList) {
+        menuItemList.clear();
+        for (IResource resource : resourceList) {                                                                  //берем список всех ресурсов
+            CheckMenuItem checkMenuItem = new CheckMenuItem(resource.getName());                                //создаем элемент меню для каждого ресурса
+            task.getResourceList()
+                    .stream()
+                    .filter(resourceOfTask -> resource.equals(resourceOfTask) && !checkMenuItem.isSelected())
+                    .forEach(resourceOfTask -> checkMenuItem.setSelected(true));
+            checkMenuItem.setOnAction(event -> {
+                if (checkMenuItem.isSelected()) {
+                    task.addResource(resource);
+                } else {
+                    task.removeResource(resource);
+                }
+            });
+            menuItemList.add(checkMenuItem);
         }
     }
 }
